requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==3.0.14
django-jsonfield==1.4.1
idna==3.4
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.7.1
urllib3==2.0.3
user-agent==0.1.10
